ostrich: Deprecate the dependency on util-eval
Problem

We want to eventually remove the dependency on util-eval in ostrich,
but it's difficult today because RuntimeEnvironment assume util-eval.

Solution

Provide a separate API which is independent of util-eval, and
deprecate the old one.

Result

We're in a position where we can remove ostrich's dependency on
util-eval.

RB_ID=605448