ostrich: Remove "start up" test in AdminServiceConfigTest Problem
AdminServiceConfigTest allocated fixed ports which doesn't
play nicely with our CI.

Solution

Instead of adding forwarders to AdminServiceConfig that could
extract the underlying HttpServer address, I opted to delete
this test for two reasons. First, it seemed like a silly test
to verify that we open a socket. Secondly, AdminServiceConfig is
deprecated in favor of AdminServiceFactory.

RB_ID=610040