csl: Publish externally with pants, sbt, and friends
Problem

It's difficult to publish util, finagle, etc. because we haven't
updated them for the new maven-less world.

Solution

Scripts that make it easy to publish with sbt and pants.  Bumped
the server-set version because we stopped publishing thrift
externally in ci.  Changed an ExitGuard test which broke under
weird sbt conditions.

Result

We're in a position where we can start publishing finagle, etc
again!

RB_ID=633570