[iago] add wily support
This adds wily support to iago. The 'finagle' victimType must be specified in the LauncherConfig for this to have any effect.

The base twitter dtab is loaded in the parrot server, and an additional dtab will be concatenated from the dtabAdd setting in the ParrotLauncherConfig which is passed through to the ParrotServerConfig. A dtab override can also be specified at runtime with '-S dtab.add=$YOUR_DTAB'.

The -S flag was added to the parrot launcher to allow arbitrary properties to be passed through to the parrot server. It may be desirable for the feeder to accept properties some day as well at some point. If a dtab is added at runtime, it is concatenated to the ParrotLauncherConfig setting. This will provide the least surprising behavior and will still allow full control over the dtab.

RB_ID=679586